+++ src/index.js
@@ -7,14 +7,22 @@ module.exports = {
   bind: (fpm) => {
     // Run When Server Init
     fpm.registerAction('INIT', () => {
-      const mqttserverOption = fpm.getConfig('mqttserver', {
-          host: 'localhost',
-          port: 1883,
-          username: 'admin',
-          password: '123123123',
-          clientId: 'fpm-client-nodejs-' + crypto.randomBytes(4).toString('hex') ,
-        });
-      const { host, port, username, password, clientId } = mqttserverOption;
+      const { MQTT_HOST, MQTT_PORT, MQTT_USER, MQTT_PASS, MQTT_CLIENTID } = fpm.getEnv();
+      const config = fpm.getConfig('mqttserver', {});
+      const option = _.assign({
+        host: 'localhost',
+        port: 1883,
+        username: 'admin',
+        password: '123123123',
+        clientId: 'fpm-client-nodejs-' + crypto.randomBytes(4).toString('hex') ,
+      }, config);
+      const { host, port, username, password, clientId } = Object.assign(option, {
+        host: MQTT_HOST || option.host,
+        port: MQTT_PORT || option.port,
+        username: MQTT_USER || option.username,
+        password: MQTT_PASS || option.password,
+        clientId: MQTT_CLIENTID || option.clientId,
+      });
 
       client = mqtt.connect(`mqtt://${host}:${port}`, { username, password, clientId });
       client.on('message', (topic, payload) => {
@@ -30,16 +38,27 @@ module.exports = {
         }else{
           args.payload = payload.toString();
         }
-        console.log(args)
         client.publish(args.topic, args.payload, { qos: args.qos || 1, retain: true});
         return 1;
       },
       subscribe: async (args) => {
-        client.subscribe(args.topic);
+        const { topic } = args;
+        if(_.isArray(topic)){
+          client.subscribe(topic);
+        }else if(_.isString(topic)){
+          const topics = topic.split(',');
+          client.subscribe(topics);
+        }
         return 1;
       },
       unsubscribe: async (args) => {
-        client.unsubscribe(args.topic);
+        const { topic } = args;
+        if(_.isArray(topic)){
+          client.unsubscribe(topic);
+        }else if(_.isString(topic)){
+          const topics = topic.split(',');
+          client.unsubscribe(topics);
+        }
         return 1;
       },
     };