+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+Support:
+- an array of topics to [un]subscribe
+- enviroment important.
+
 ## 0.0.2
 Add option param format:hex/string for mqttclient.publish